Verify user selected coordinates with database
First, I set up a firebase database and then connected it via the
Scene component. A function is defined outside of above the Scene
function to define the connection to the DB. Then that function is used
in a function that handles the character selection click event. In this
selection function, the user-click coordinates and the DB coordinates
are compared.

Additionally, the Scene rendering needed the character
selection images to be rendered as a sibling of the scene-btn instead of
being a child. This way the targeting box is not set again when the user
selects a character.