Fix all characters found on second game bug
When the user starts a second game, the GameScene would show no
characters have been found, but upon finding the first character the
SubmitScoreScreen would immediately appear. The charactersFound state
was added as a dependency in the changeCharactersFound callBack to fix
this.

Also, the character found property and time state changes were separated
from the screen transition by placing them in a useEffect hook that
monitors the transition to the Leaderboard.